DungNA fix bugs

+++ Source/JPA_Spring/fparking/src/main/java/com/tagroup/fparking/service/domain/Admin.java
@@ -1,7 +1,12 @@
 package com.tagroup.fparking.service.domain;
 
 import java.io.Serializable;
-import javax.persistence.*;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.NamedQuery;
 
 
 /**
@@ -15,7 +20,7 @@ public class Admin implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	private String email;
 
@@ -32,11 +37,11 @@ public class Admin implements Serializable {
 	public Admin() {
 	}
 
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
